feat(models): add AsRef implementation for Footer
This commit adds an implementation of the AsRef trait for the Footer struct. This allows a Footer instance to be referenced as a string slice, facilitating easier manipulation and comparison of Footer objects in the codebase.

NEW FEATURE: You appear to have introduced one or more new features that are backward-compatible. According to Semantic Versioning (SemVer), this requires a minor version update. Please verify and update your version number accordingly.